Correção de erros e finalização da aplicação.

+++ CL.Data/Repository/ClienteRepository.cs
@@ -1,11 +1,8 @@
 ﻿using CL.Core.Domain;
 using CL.Data.Context;
-using CL.Manager.Interfaces;
+using CL.Manager.Interfaces.Repositories;
 using Microsoft.EntityFrameworkCore;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace CL.Data.Repository
@@ -13,31 +10,28 @@ namespace CL.Data.Repository
     public class ClienteRepository : IClienteRepository
     {
         private readonly ClContext context;
+
         public ClienteRepository(ClContext context)
         {
             this.context = context;
         }
 
-        public async Task<IEnumerable<Cliente>> GetClientesAsync() // Método para retornar todos da tabela de cliente
+        public async Task<IEnumerable<Cliente>> GetClientesAsync()
         {
             return await context.Clientes
                 .Include(p => p.Endereco)
                 .Include(p => p.Telefones)
-                .AsNoTracking().ToListAsync(); // AsNoTracking = ganho de performance
+                .AsNoTracking().ToListAsync();
         }
 
-        public async Task<Cliente> GetClienteAsync(int Id) // Método de retorno de clientes por Id
+        public async Task<Cliente> GetClienteAsync(int id)
         {
-            //return await context.Clientes.Where(c => c.Id_Cliente == Id_Cliente).FirstOrDefaultAsync();
-            //return await context.Clientes.AsNoTracking().FirstOrDefaultAsync(c => c.Id_Cliente == Id_Cliente);
-            //return await context.Clientes.SingleOrDefaultAsync(c => c.Id_Cliente == Id_Cliente);
             return await context.Clientes
                 .Include(p => p.Endereco)
                 .Include(p => p.Telefones)
-                .SingleOrDefaultAsync(p => p.ClienteId == Id);
+                .SingleOrDefaultAsync(p => p.Id == id);
         }
 
-        //Insert
         public async Task<Cliente> InsertClienteAsync(Cliente cliente)
         {
             await context.Clientes.AddAsync(cliente);
@@ -45,27 +39,42 @@ namespace CL.Data.Repository
             return cliente;
         }
 
-        //Update
         public async Task<Cliente> UpdateClienteAsync(Cliente cliente)
         {
-            var clienteConsultado = await context.Clientes.FindAsync(cliente.ClienteId);
+            var clienteConsultado = await context.Clientes
+                                                 .Include(p => p.Endereco)
+                                                 .Include(p => p.Telefones)
+                                                 .FirstOrDefaultAsync(p => p.Id == cliente.Id);
             if (clienteConsultado == null)
             {
                 return null;
             }
-            //clienteConsultado.Nome_Cliente = cliente.Nome_Cliente;
-            //clienteConsultado.DataNascimento = cliente.DataNascimento;
-            context.Entry(clienteConsultado).CurrentValues.SetValues(cliente); //Verifica todas as propriedades
+            context.Entry(clienteConsultado).CurrentValues.SetValues(cliente);
+            clienteConsultado.Endereco = cliente.Endereco;
+            UpdateClienteTelefones(cliente, clienteConsultado);
             await context.SaveChangesAsync();
             return clienteConsultado;
         }
 
-        //Delete
-        public async Task DeleteClienteAsync(int Id) //Não usa Task<Cliente> pq não tem que retornar nada
+        private void UpdateClienteTelefones(Cliente cliente, Cliente clienteConsultado)
         {
-            var clienteConsultado = await context.Clientes.FindAsync(Id);
-            context.Clientes.Remove(clienteConsultado);
+            clienteConsultado.Telefones.Clear();
+            foreach (var telefone in cliente.Telefones)
+            {
+                clienteConsultado.Telefones.Add(telefone);
+            }
+        }
+
+        public async Task<Cliente> DeleteClienteAsync(int id)
+        {
+            var clienteConsultado = await context.Clientes.FindAsync(id);
+            if (clienteConsultado == null)
+            {
+                return null;
+            }
+            var clienteRemovido = context.Clientes.Remove(clienteConsultado);
             await context.SaveChangesAsync();
+            return clienteRemovido.Entity;
         }
     }
 }
